requirements/base.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
Django==4.1.5
django-bootstrap5==23.1
django-widget-tweaks==1.4.8
pytz==2022.7.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
django-environ
Pillow
django_cleanup
django-imagekit
opencv-python